AudioExtractorAgent/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
groq
pydub
